Ajustes em Conversão
Converte de Binario para Decimal, e Converte com sinal negativo de
Decimal Binario, problemas com conversão com numeros menores que 1 e
maiores que 0.

+++ Conversao/conversao.h
@@ -11,8 +11,8 @@ using namespace std;
 
 namespace Conv
 {
-    string DecBin(double decVal, int maxRep);//Converte de Decimal para Binario.
-    int getExp(double decVal, int maxRep);
+    string DecBin(float decVal, int maxRep);//Converte de Decimal para Binario.
+    int getExp(float decVal, int maxRep);
     string DecBin(int Dec, int Inf, int Sup);//Converte de Decimal para Binario.
     string BinDec(string binInt, string binDec);//Converte de Binario para Decimal.
 }
